fix some issues
- add a test server implementation

This test server is only meant for testing use! Under no circumstances should it ever be used in a production environment. It's also incomplete and lacks handling for most packages.
It is currently able to complete the handshake process, but does not store any information about the player.

While it is also possible to test this server without a minecraft client running (by enabling it in the source file) this is heavily discouraged.